refactor(backends): keep FirecrackerBackend a real backend, free FcDriver of it
A VmBackend that cannot start a VM is not a backend. Moving the shell
into mvm-backends had cost it its start path — fc_runner() lives in
mvm-runtime — and I had papered over that with a refusal. Inheriting the
trait's default refusal instead would only have made the same lie tidier.

The shell goes back to mvm-runtime, beside the runner it starts through,
with fc_runner().start(config) intact. It is a genuine VmBackend again,
and every method the trait promises works.

FcDriver stops wrapping it. It was reaching through the shell for five
facts — name, kind, is_available, security_profile, guest_channel_info —
all of which are the driver's own to report: two constants, one call to
fc::host::is_installed(), the claim-bearing Tier 1 profile, and a refusal
for a guest channel Firecracker does not expose. That removes a
driver-to-legacy-shell dependency rather than relocating one, and drops
the VmBackend import from the driver entirely.

Two tests changed meaning rather than being deleted: the capability test
now asserts the driver's own tier, and the guest-channel test asserts the
refusal instead of delegation to a shell the driver no longer holds.
